+++ include/MathUtil.h
@@ -4,7 +4,7 @@
 
 namespace math{
 #define EPSILON 0.00001f
-inline bool equal(float d1, float d2) {
+inline bool equal(double d1, double d2) {
     return  std::abs(d1 - d2) < EPSILON;
 }
 }
