+++ include/Tuple.h
@@ -11,9 +11,9 @@ using namespace math;
 
 class Tuple : public VectorTrait<Tuple>{
 private:
-    std::array<float, 4> m_data{};
+    std::array<double, 4> m_data{};
 public:
-    Tuple(float x, float y, float z, float w) {
+    Tuple(double x, double y, double z, double w) {
         m_data[0] = x;
         m_data[1] = y;
         m_data[2] = z;
@@ -33,22 +33,22 @@ public:
             && equal(m_data[3], m_data[3]);
     }
 
-    float& operator[](size_t index) {
+    double& operator[](size_t index) {
         return m_data[index];
     }
-    const float operator[](size_t index) const {
+    const double operator[](size_t index) const {
         return m_data[index];
     }
 
-    const std::array<float, 4>& getData() const {return m_data;}    
+    const std::array<double, 4>& getData() const {return m_data;}    
     
     bool IsPoint() const;
     bool IsVector() const;
     
     Tuple Reflect(const Tuple normal) const;
 
-    static Tuple Point(float x, float y, float z);
-    static Tuple Vector(float x, float y, float z);
+    static Tuple Point(double x, double y, double z);
+    static Tuple Vector(double x, double y, double z);
 
     friend std::ostream& operator<<(std::ostream& os, const Tuple& t);
 };
